requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap5==2024.10
dj-database-url==2.3.0
dj-email-url==1.0.6
Django==5.1.2
django-allauth==65.1.0
django-cache-url==3.4.5
django-crispy-forms==2.3
django-jalali-date==1.1.3
django-rosetta==0.10.1
environs==11.0.0
idna==3.10
jalali_core==1.0.0
jdatetime==5.0.0
marshmallow==3.23.0
packaging==24.1
pillow==11.0.0
polib==1.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.6.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
